Add read operation to perfuser
Reading the file descriptor returns values of per-thread counters related to
perfuser state, namely:

- nmi: NMI count
- irq: IRQ count (should be equals to NMI)
- sig: successfull send_sig_info()
- blk: number of blocked, possibly delayed signals
- err: errors related to memory allocation, means results are unreliable
- ts: perf timestamps, if available (not implemented yet)

The read verify that the userspace structure has the same size as the kernel
one.

The values monotonically increase until an overflow occurs. A perf sample can
occur while the read is in progress. Really, atomic values are not needed. We
should instead use two structures, and swap structure pointers when update is
completed.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>